+++ ci4/app/Interfaces/RoleRule.php
@@ -1,0 +1,9 @@
+<?php namespace App\Interfaces;
+/**
+ * @property string $apiGroup
+ * @property string $resource
+ * @property string $verbs # comma-separated
+ */
+interface RoleRule {
+
+}