+++ ci4/app/Interfaces/RoleRuleList.php
@@ -1,0 +1,7 @@
+<?php namespace App\Interfaces;
+/**
+ * @property RoleRule[] $values
+ */
+interface RoleRuleList {
+
+}