+++ ci4/app/Libraries/DeploymentSteps/RoleStep.php
@@ -1,24 +1,23 @@
 <?php namespace App\Libraries\DeploymentSteps;
 
 use App\Entities\Deployment;
-use App\Entities\DeploymentSpecificationClusterRoleRule;
+use App\Entities\DeploymentSpecificationRoleRule;
 use App\Libraries\DeploymentSteps\Helpers\DeploymentStepHelper;
 use App\Libraries\DeploymentSteps\Helpers\DeploymentSteps;
 use App\Libraries\Kubernetes\KubeAuth;
-use App\Models\DeploymentSpecificationClusterRoleRuleModel;
+use App\Models\DeploymentSpecificationRoleRuleModel;
 use RenokiCo\PhpK8s\Exceptions\KubernetesAPIException;
-use RenokiCo\PhpK8s\Instances\Subject;
-use RenokiCo\PhpK8s\Kinds\K8sClusterRole;
-use RenokiCo\PhpK8s\Kinds\K8sClusterRoleBinding;
+use RenokiCo\PhpK8s\Instances\Rule;
+use RenokiCo\PhpK8s\Kinds\K8sRole;
 
-class ClusterRoleBindingStep extends BaseDeploymentStep {
+class RoleStep extends BaseDeploymentStep {
 
     public function getIdentifier(): string {
-        return DeploymentSteps::ClusterRoleBinding;
+        return DeploymentSteps::Role;
     }
 
     public function getName(): string {
-        return 'Cluster Role Binding';
+        return 'Role';
     }
 
     public function hasPreviewCommand(): bool {
@@ -47,15 +46,15 @@ class ClusterRoleBindingStep extends BaseDeploymentStep {
 
     public function getSuccessStatus(Deployment $deployment): string {
         $spec = $deployment->findDeploymentSpecification();
-        /** @var DeploymentSpecificationClusterRoleRule $hasClusterRoleRules */
-        $hasClusterRoleRules = (new DeploymentSpecificationClusterRoleRuleModel())
+        /** @var DeploymentSpecificationRoleRule $hasRoleRules */
+        $hasRoleRules = (new DeploymentSpecificationRoleRuleModel())
             ->where('deployment_specification_id', $spec->id)
             ->find();
-        $expectResource = $hasClusterRoleRules->exists();
+        $expectResource = $hasRoleRules->exists();
 
         return $expectResource
-            ? DeploymentStepHelper::ClusterRoleBinding_Found
-            : DeploymentStepHelper::ClusterRoleBinding_NotFoundNotExpected;
+            ? DeploymentStepHelper::Role_Found
+            : DeploymentStepHelper::Role_NotFoundNotExpected;
     }
 
     /**
@@ -89,21 +88,21 @@ class ClusterRoleBindingStep extends BaseDeploymentStep {
         $resource = $this->getResource($deployment, true);
 
         $spec = $deployment->findDeploymentSpecification();
-        /** @var DeploymentSpecificationClusterRoleRule $hasClusterRoleRules */
-        $hasClusterRoleRules = (new DeploymentSpecificationClusterRoleRuleModel())
+        /** @var DeploymentSpecificationRoleRule $hasRoleRules */
+        $hasRoleRules = (new DeploymentSpecificationRoleRuleModel())
             ->where('deployment_specification_id', $spec->id)
             ->find();
-        $expectResource = $hasClusterRoleRules->exists();
+        $expectResource = $hasRoleRules->exists();
 
         $hasAlias = $resource->exists();
         if ($hasAlias) {
             return $expectResource
-                ? DeploymentStepHelper::ClusterRoleBinding_Found
-                : DeploymentStepHelper::ClusterRoleBinding_FoundNotExpected;
+                ? DeploymentStepHelper::Role_Found
+                : DeploymentStepHelper::Role_FoundNotExpected;
         } else {
             return $expectResource
-                ? DeploymentStepHelper::ClusterRoleBinding_NotFound
-                : DeploymentStepHelper::ClusterRoleBinding_NotFoundNotExpected;
+                ? DeploymentStepHelper::Role_NotFound
+                : DeploymentStepHelper::Role_NotFoundNotExpected;
         }
     }
 
@@ -111,22 +110,7 @@ class ClusterRoleBindingStep extends BaseDeploymentStep {
         if (strlen($deployment->name) == 0) {
             return 'Missing name';
         }
-        if (strlen($deployment->namespace) == 0) {
-            return 'Missing namespace';
-        }
 
-        $namespaceStep = new NamespaceStep();
-        if ($namespaceStep->getStatus($deployment) != DeploymentStepHelper::Namespace_Found) {
-            return 'Missing Namespace';
-        }
-        $clusterRoleStep = new ClusterRoleStep();
-        if ($clusterRoleStep->getStatus($deployment) != $clusterRoleStep->getSuccessStatus($deployment)) {
-            return 'Missing Cluster Role';
-        }
-        $serviceAccount = new ServiceAccountStep();
-        if ($serviceAccount->getStatus($deployment) != DeploymentStepHelper::ServiceAccount_Found) {
-            return 'Missing Service Account';
-        }
         return null;
     }
 
@@ -152,20 +136,24 @@ class ClusterRoleBindingStep extends BaseDeploymentStep {
     /**
      * @throws \Exception
      */
-    private function getResource(Deployment $deployment, bool $auth = false): K8sClusterRoleBinding {
-        $resource = new K8sClusterRoleBinding();
+    private function getResource(Deployment $deployment, bool $auth = false): K8sRole {
+        $spec = $deployment->findDeploymentSpecification();
+
+        $resource = new K8sRole();
         $resource
-            ->setName("{$deployment->name}.{$deployment->namespace}")
-            ->addSubject(
-                (new Subject())
-                    ->setAttribute('kind', 'ServiceAccount')
-                    ->setAttribute('name', $deployment->name)
-                    ->setAttribute('namespace', $deployment->namespace)
-            )
-            ->setRole(
-                (new K8sClusterRole())
-                    ->setName("{$deployment->name}.{$deployment->namespace}")
-            );
+            ->setName($deployment->name)
+            ->setNamespace($deployment->namespace);
+
+        $rules = [];
+        $spec->deployment_specification_role_rules->find();
+        foreach ($spec->deployment_specification_role_rules as $roleRule) {
+            $rule = new Rule();
+            $rule->addApiGroup($roleRule->api_group);
+            $rule->addResource($roleRule->resource);
+            $rule->addVerbs(explode(',', $roleRule->verbs));
+            $rules[] = $rule;
+        }
+        $resource->setRules($rules);
 
         if ($auth) {
             $auth = new KubeAuth();
@@ -174,5 +162,4 @@ class ClusterRoleBindingStep extends BaseDeploymentStep {
 
         return $resource;
     }
-
 }