+++ vue/src/core/services/Deploy/models/DeploymentSpecificationRoleRule.ts
@@ -1,0 +1,14 @@
+/**
+ * Created by ModelParser
+ * Date: 13-12-2024.
+ * Time: 15:21.
+ */
+import {DeploymentSpecificationRoleRuleDefinition} from './definitions/DeploymentSpecificationRoleRuleDefinition';
+
+export class DeploymentSpecificationRoleRule extends DeploymentSpecificationRoleRuleDefinition {
+
+    constructor(json?: any) {
+        super(json);
+    }
+
+}